requirements.txt
fastapi
uvicorn
requests
Pillow
numpy
tensorflow
pydantic